trying to fix document in tests
